+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add sleep of 100ms to XMPP output to ensure message is sent before connection closes
 
 ## [0.2.0] - 2019-05-16
 ### Changed